+++ lib/enum/dark_option.dart
@@ -1,2 +1,2 @@
-// fixed
+
 enum DarkOption { dynamic, alwaysOn, alwaysOff }