+++ lib/theme.dart
@@ -1,6 +1,5 @@
 library theme;
 
-// fixed
 export 'app_theme.dart';
 export 'enum/theme_types.dart';
 export 'models/theme_model.dart';